fix: UI refactoring = components, types, hooks, utils
